fixed some gross bugs

+++ src/TM1638lite.h
@@ -18,7 +18,7 @@
 class TM1638lite
 {
   public:
-    TM1638lite(int strobe, int clock, int data);
+    TM1638lite(uint8_t strobe, uint8_t clock, uint8_t data);
     void sendCommand(uint8_t value);
     void reset();
     uint8_t readButtons();
